+++ kubejs/server_scripts/mod_scripts/applied_energistics/inscriber.js
@@ -1,0 +1,8 @@
+const inscriber = [
+    {
+        output: { 'item': "appliedenergistics2:printed_silicon" },
+        top: { 'item': "appliedenergistics2:silicon_press" },
+        middle: { 'tag': "refinedstorage:silicon" },
+        ID: 'appliedenergistics2:inscriber/silicon_print'
+    }
+]