+++ kubejs/server_scripts/mod_scripts/powah/energizing.js
@@ -1,0 +1,9 @@
+
+const energizing = [
+    {
+        inputs: [{ "item": "elementalcraft:fire_shard" }, { "item": "elementalcraft:air_shard" }, { "item": "elementalcraft:water_shard" }, { "item": "elementalcraft:earth_shard" }],
+        energy: 10000,
+        output: { "item": 'elementalcraft:inertcrystal', 'count': 8 },
+        ID: 'waddlecraft:inert_crystal'
+    }
+]