+++ kubejs/server_scripts/mod_scripts/thermal/pulverizing.js
@@ -8,5 +8,15 @@ const thermalpulverizer = [
         output: Item.of('industrialforegoing:dryrubber').withChance(2.5),
         input: 'rats:plastic_waste',
         energy: '4000'
+    },
+    {
+        output: Item.of('immersiveengineering:dust_coke').withChance(1),
+        input: 'immersiveengineering:coal_coke',
+        energy: '2000'
+    },
+    {
+        output: Item.of('immersivepetroleum:petcoke_dust').withChance(1),
+        input: 'immersivepetroleum:petcoke',
+        energy: '2000'
     }
 ]