+++ kubejs/server_scripts/vanilla_utilities/crafting_table/shapelesscrafting.js
@@ -20,5 +20,20 @@ const shapelessRecipes = [
 		output: 'gobber2:gobber2_seed_end',
 		input: ['gobber2:gobber2_seed_nether', '4x bigreactors:magentite_ingot', '4x upgradednetherite_ultimate:ultimate_upgraded_netherite_ingot'],
 		id: 'gobber2:gobber2_seed_end'
+	},
+	{
+		output: 'minecraft:trapped_chest',
+		input: ['#forge:chests/trapped'],
+		id: 'minecraft:trapped_chest'
+	},
+	{
+		output: '2x silentgear:azure_silver_dust',
+		input: ['silentgear:crushed_shulker_shell', 'thermal:enderium_dust', 'thermal:silver_dust'],
+		id: 'waddlecraft:azure_silver_dust'
+	},
+	{
+		output: '2x silentgear:crimson_iron_dust',
+		input: ['minecraft:nether_wart', 'thermal:iron_dust', 'minecraft:blaze_powder'],
+		id: 'waddlecraft:azure_silver_dust'
 	}
 ]